tetris: break up into chunks suitable for book
