+++ src/tetris/main.cpp
@@ -74,118 +74,10 @@ public:
     }while(!gameover_);
   }
 
-  // returns true if the game has ended as a result of this move down
-  bool MoveDown() {
-    const std::lock_guard<std::mutex> lock(mtx_);
-    int y, x;
-    if(PrepForMove(&y, &x)){
-      if(PieceStuck()){
-        if(y <= board_top_y_ - 2){
-          return true;
-        }
-        curpiece_->mergedown(*board_);
-        curpiece_ = NewPiece();
-      }else{
-        ++y;
-        if(!curpiece_->move(y, x) || !nc_.render()){
-          throw TetrisNotcursesErr("move() or render()");
-        }
-      }
-    }
-    return false;
-  }
-
-  void MoveLeft() {
-    const std::lock_guard<std::mutex> lock(mtx_);
-    int y, x;
-    if(!PrepForMove(&y, &x)){
-      return;
-    }
-    // For each line of the current piece, find the leftmost populated column.
-    // Check the game area to the immediate left. If something's there, we
-    // can't make this move.
-    ncpp::Cell c;
-    for(int ly = 0 ; ly < curpiece_->get_dim_y() ; ++ly){
-      int lx = 0;
-      while(lx < curpiece_->get_dim_x()){
-        if(curpiece_->get_at(ly, lx, &c)){
-          if(c.get().gcluster && c.get().gcluster != ' '){
-            break;
-          }
-        }
-        ++lx;
-      }
-      if(lx < curpiece_->get_dim_x()){ // otherwise, nothing on this row
-        ncpp::Cell b;
-        int cmpy = ly, cmpx = lx - 1;
-        curpiece_->translate(*board_, &cmpy, &cmpx);
-        if(board_->get_at(cmpy, cmpx, &b)){
-          if(b.get().gcluster && b.get().gcluster != ' '){
-            return; // move is blocked
-          }
-        }
-      }
-    }
-    --x;
-    if(!curpiece_->move(y, x) || !nc_.render()){ // FIXME needs y?
-      throw TetrisNotcursesErr("move() or render()");
-    }
-  }
-
-  void MoveRight() {
-    const std::lock_guard<std::mutex> lock(mtx_);
-    int y, x;
-    if(!PrepForMove(&y, &x)){
-      return;
-    }
-    // For each line of the current piece, find the rightmost populated column.
-    // Check the game area to the immediate right. If something's there, we
-    // can't make this move.
-    ncpp::Cell c;
-    for(int ly = 0 ; ly < curpiece_->get_dim_y() ; ++ly){
-      int lx = curpiece_->get_dim_x() - 1;
-      while(lx >= 0){
-        if(curpiece_->get_at(ly, lx, &c)){
-          if(c.get().gcluster && c.get().gcluster != ' '){
-            break;
-          }
-        }
-        --lx;
-      }
-      if(lx >= 0){ // otherwise, nothing on this row
-        ncpp::Cell b;
-        int cmpy = ly, cmpx = lx + 1;
-        curpiece_->translate(*board_, &cmpy, &cmpx);
-        if(board_->get_at(cmpy, cmpx, &b)){
-          if(b.get().gcluster && b.get().gcluster != ' '){
-            return; // move is blocked
-          }
-        }
-      }
-    }
-    ++x;
-    if(!curpiece_->move(y, x) || !nc_.render()){ // FIXME needs y?
-      throw TetrisNotcursesErr("move() or render()");
-    }
-  }
-
-  void RotateCcw() {
-    const std::lock_guard<std::mutex> lock(mtx_);
-    int y, x;
-    if(!PrepForMove(&y, &x)){
-      return;
-    }
-    // FIXME rotate that fucker ccw
-  }
-
-  void RotateCw() {
-    const std::lock_guard<std::mutex> lock(mtx_);
-    int y, x;
-    if(!PrepForMove(&y, &x)){
-      return;
-    }
-    // FIXME rotate that fucker cw
-  }
+#include "movedown.h"
+#include "moveleft.h"
+#include "moveright.h"
+#include "rotate.h"
 
 private:
   ncpp::NotCurses& nc_;
@@ -210,60 +102,8 @@ private:
   }
 
 #include "background.h"
-
-  bool PieceStuck() {
-    if(!curpiece_){
-      return false;
-    }
-    // check for impact. iterate over bottom row of piece's plane, checking for
-    // presence of glyph. if there, check row below. if row below is occupied,
-    // we're stuck.
-    int y, x;
-    curpiece_->get_dim(&y, &x);
-    --y;
-    while(x--){
-      int cmpy = y + 1, cmpx = x; // need game area coordinates via translation
-      curpiece_->translate(*board_, &cmpy, &cmpx);
-      ncpp::Cell c;
-      if(board_->get_at(cmpy, cmpx, &c) < 0){
-        throw TetrisNotcursesErr("get_at()");
-      }
-      if(c.get().gcluster && c.get().gcluster != ' '){
-        return true;
-      }
-    }
-    return false;
-  }
-
-  // tidx is an index into tetriminos. yoff and xoff are relative to the
-  // terminal's origin. returns colored north-facing tetrimino on a plane.
-  std::unique_ptr<ncpp::Plane> NewPiece() {
-    const int tidx = random() % 7;
-    const struct tetrimino* t = &tetriminos[tidx];
-    const size_t cols = strlen(t->texture);
-    int y, x;
-    stdplane_->get_dim(&y, &x);
-    const int xoff = x / 2 - BOARD_WIDTH + (random() % BOARD_WIDTH - 1);
-    std::unique_ptr<ncpp::Plane> n = std::make_unique<ncpp::Plane>(2, cols, board_top_y_ - 2, xoff, nullptr);
-    if(n){
-      uint64_t channels = 0;
-      channels_set_bg_alpha(&channels, CELL_ALPHA_TRANSPARENT);
-      channels_set_fg_alpha(&channels, CELL_ALPHA_TRANSPARENT);
-      n->set_fg(t->color);
-      n->set_bg_alpha(CELL_ALPHA_TRANSPARENT);
-      n->set_base(channels, 0, "");
-      y = 0;
-      for(size_t i = 0 ; i < strlen(t->texture) ; ++i){
-        if(t->texture[i] == '*'){
-          if(n->putstr(y, x, "██") < 0){
-            return NULL;
-          }
-        }
-        y += ((x = ((x + 2) % cols)) == 0);
-      }
-    }
-    return n;
-  }
+#include "stuck.h"
+#include "newpiece.h"
 
 };
